feat: use an url hash to share ids
