+++ app/[compositeKey]/page.tsx
@@ -1,5 +1,7 @@
-import { Client } from "../unseal/client";
+import { redirect } from "next/navigation";
 
+// This page is here for backwards compatibility with old links.
+// Old links were of the form /{compositeKey} and now they are of the form /unseal#{compositeKey}
 export default function Page(props: { params: { compositeKey: string } }) {
-  return <Client compositeKey={props.params.compositeKey} />;
+  return redirect(`/unseal#${props.params.compositeKey}`);
 }
